requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.11.12
click==8.3.1
distro==1.9.0
fastapi==0.122.0
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
jiter==0.12.0
numpy==2.3.5
openai==2.8.1
portalocker==3.2.0
protobuf==6.33.1
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
qdrant-client==1.16.1
sniffio==1.3.1
starlette==0.50.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
